requirements.txt
discord.py
discord-py-slash-command
requests
aiohttp
python-dotenv